feat #25 에러 처리

+++ src/main/java/dgu/cse/newsee/service/category/CategoryService.java
@@ -1,7 +1,9 @@
 package dgu.cse.newsee.service.category;
 
+import dgu.cse.newsee.apiPayload.exception.UserException;
 import dgu.cse.newsee.domain.enums.Category;
 import dgu.cse.newsee.repository.UserCategoryRepository;
+import dgu.cse.newsee.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,10 +14,12 @@ import java.util.List;
 public class CategoryService {
 
     private final UserCategoryRepository userCategoryRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public CategoryService(UserCategoryRepository userCategoryRepository) {
+    public CategoryService(UserCategoryRepository userCategoryRepository, UserRepository userRepository) {
         this.userCategoryRepository = userCategoryRepository;
+        this.userRepository = userRepository;
     }
 
     // 모든 관심분야 가져오기
@@ -25,11 +29,17 @@ public class CategoryService {
 
     // 사용자의 관심분야 가져오기
     public List<Category> getUserCategories(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserException.UserNonExistsException("등록되지 않은 사용자입니다.");
+        }
         return userCategoryRepository.findCategoriesByUserId(userId);
     }
 
     // 사용자의 관심분야 수정하기
     public void updateUserCategories(Long userId, List<Category> categories) {
+        if (!userRepository.existsById(userId)) {
+            throw new UserException.UserNonExistsException("등록되지 않은 사용자입니다.");
+        }
         // 기존 관심 분야 삭제
         userCategoryRepository.deleteByUserId(userId);
